requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
aiosqlite==0.20.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
